moved googlekeep related models to `pkg`
